refactor: rename IHttpProxy -> IHttpClient
