+++ src/lib/interfaces/IHttpClient.ts
@@ -6,7 +6,7 @@ export interface ResponseHeaders {
 	[other: string]: unknown;
 }
 
-export interface IHttpProxy {
+export interface IHttpClient {
 	get(url: string, headers?: RequestHeaders, options?: Record<string, unknown>): Promise<{ status: number, headers: ResponseHeaders, data: unknown }>;
 	post(url: string, body: any, headers?: RequestHeaders): Promise<{ status: number, headers: ResponseHeaders, data: unknown }>;
 }
